% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoform_links_database-class.R
\name{add3pEndDatabase,IsoformDatabase-method}
\alias{add3pEndDatabase,IsoformDatabase-method}
\title{Title}
\usage{
\S4method{add3pEndDatabase}{IsoformDatabase}(x, custom_promoter_annotation, reference_annotation, window)
}
\arguments{
\item{IsoformDatabase}{}
}
\description{
Title
}
