% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoform_links_database-class.R
\name{addPromoterDatabase,IsoformDatabase-method}
\alias{addPromoterDatabase,IsoformDatabase-method}
\title{Title}
\usage{
\S4method{addPromoterDatabase}{IsoformDatabase}(
  x,
  custom_promoter_annotation,
  reference_annotation,
  window
)
}
\arguments{
\item{IsoformDatabase}{}
}
\description{
Title
}
