% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{avg_update}
\alias{avg_update}
\title{Incremental average calculation}
\usage{
avg_update(avg, cnt, val, na.rm = TRUE)
}
\arguments{
\item{avg}{The current average value}

\item{cnt}{The count of elements before the new value}

\item{val}{The new value}

\item{na.rm}{Whether to remove NA values}
}
\value{
The updated average
}
\description{
Online algorithm for updating a given average on new value
TODO: check for numerical stability
}
