% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_load}
\alias{cache_load}
\title{Load cache for a specific range}
\usage{
cache_load(range, data_path = "data/cache", tc = "hour")
}
\arguments{
\item{range}{The range of the cache to load}

\item{data_path}{The path to look for cache files}

\item{tc}{The date unit used to split cache files (granularity)}
}
\value{
data.table object
}
\description{
Load cache for a specific range
}
