% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_purge_expired}
\alias{cache_purge_expired}
\title{Remove cache data older than ttl seconds}
\usage{
cache_purge_expired(ttl = 2592000, data_path = "data/cache",
  archive_path = NULL, remove_empty = TRUE)
}
\arguments{
\item{ttl}{ttl in seconds}

\item{data_path}{Where to look for expired cache entries}

\item{archive_path}{A path where to copy expired cache files (for archiving)}

\item{remove_empty}{Whether to remove empty directories at the end}
}
\value{
invisible(NULL)
}
\description{
Remove cache data older than ttl seconds
}
