% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_state_range}
\alias{cache_state_range}
\title{Get cache state for a specific single or multi range}
\usage{
cache_state_range(range, tc = "hour", data_path = "data/cache")
}
\arguments{
\item{range}{A single range-list or list of range-lists in the format list(gte = ..., lte = ...) expressing the date ranges to check in cache}

\item{tc}{The sampling period of the cache}

\item{data_path}{The data path to look at}
}
\value{
A list with hits and misses for files and dates
}
\description{
Get cache state for a specific single or multi range
}
\examples{
r = range_recent_tc(what = "seasonal", tc = "hour", N = 30, sc = "day")
cache_state_H  <- cache_state_range(range = r, tc = "hour", data_path = "data/cache")

}
