% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_state_range_recent}
\alias{cache_state_range_recent}
\title{Get cache state for a specific continuous range (recent data)}
\usage{
cache_state_range_recent(range, tc = "hour", data_path = "data/cache")
}
\arguments{
\item{range}{A list in the format list(gte = ..., lte = ...) expressing the date range to check in cache}

\item{tc}{The sampling period of the cache}

\item{data_path}{The data path to look at}
}
\value{
A list with hits and misses for files and dates
}
\description{
Get cache state for a specific continuous range (recent data)
}
\examples{
cache_state_24h  <- cache_state_range_recent(range = range_recent_tc(N = 24),
                                             tc = "hour", data_path = "data/cache")

}
