% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_state_range_seasonal}
\alias{cache_state_range_seasonal}
\title{Get cache state for a specific seasonal range}
\usage{
cache_state_range_seasonal(range, data_path = "data/cache")
}
\arguments{
\item{range}{A list of seasonal range-lists in the format list(gte = ..., lte = ...) expressing the date ranges to check in cache}

\item{data_path}{The data path to look at}
}
\value{
A list with hits and misses for files and dates
}
\description{
Get cache state for a specific seasonal range
}
\examples{
cache_state_24h  <- cache_state_range_seasonal(range = range_recent_tc(what = "seasonal", N = 24),
                                               tc = "hour", data_path = "data/cache")

}
