% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePromoterDominance.R
\name{calculatePromoterDominance}
\alias{calculatePromoterDominance}
\title{Caculate promoter dominance}
\usage{
calculatePromoterDominance(LATER, IsoformDatabase)
}
\arguments{
\item{countData}{counts data from 5'-3' links}

\item{pairsDataBase}{database of 5'-3' link isoforms}
}
\description{
Caculate promoter dominance
}
