% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countLinks.R
\name{countLinks}
\alias{countLinks}
\title{count full length reads to 5'-'3 isoforms}
\usage{
countLinks(alignments, IsoformDatabase)
}
\arguments{
\item{alignments}{genomic ranges object of alignments (samtools bam to bed)}

\item{linksDatabase}{5'-3' isoform link database}
}
\description{
count full length reads to 5'-'3 isoforms
}
