% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signals.R
\name{create_handler_registration}
\alias{create_handler_registration}
\title{Create a New Handler Registration}
\usage{
create_handler_registration()
}
\description{
Create a new handler registration. Objects generated from this function can
implement the interface described in \code{\link{events}} -- handlers can be
attached and detached with \code{on} / \code{off}. and events can be emitted
with \code{emit}.
}

