% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defer.R
\name{defer}
\alias{defer}
\title{Defer Evaluation of an Expression}
\usage{
defer(expr, envir = parent.frame(), priority = c("first", "last"))
}
\arguments{
\item{expr}{An expression to be evaluated.}

\item{envir}{Attach exit handlers to this environment.
Typically, this should be either the current environment or
a parent frame (accessed through \code{\link{parent.frame}()}).}

\item{priority}{Specify whether if this handler should be executed
\code{"first"} or \code{"last"}, relative to any other registered
handlers on this environment.}
}
\description{
Similar to \code{\link{on.exit}()}, but allows one to attach
an expression to be evaluated when exitting any frame currently
on the stack.
}

