% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{dir.hasFiles}
\alias{dir.hasFiles}
\title{Check if directory has files}
\usage{
dir.hasFiles(d, recursive = TRUE)
}
\arguments{
\item{d}{The directory to check}

\item{recursive}{Whether to check recursively or just in the first level}
}
\value{
Logical (TRUE) if directory d has any file
}
\description{
Check if the directory has any files recursively
}
