% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{dir.isLeaf}
\alias{dir.isLeaf}
\title{Check if directory is leaf directory}
\usage{
dir.isLeaf(d)
}
\arguments{
\item{d}{The directory to check}
}
\value{
Logical (TRUE) if directory d is a leaf
}
\description{
Check if directory is a leaf directory, i.e. has not descendant directories
}
