% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{dir.removeEmpty}
\alias{dir.removeEmpty}
\title{Remove empty directories}
\usage{
dir.removeEmpty(dpath)
}
\arguments{
\item{dpath}{The path under which this function should look for empty directories}
}
\description{
Remove empty directories under a specific path
}
