% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatePromoterDominance.R
\name{estimatePromoterDominance}
\alias{estimatePromoterDominance}
\title{estimatePromoterDominance}
\usage{
estimatePromoterDominance(LATER, IsoformDatabase, method)
}
\arguments{
\item{method}{}
}
\description{
estimatePromoterDominance
}
