% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateTranscriptionalBias.R
\name{estimateTranscriptionalBias}
\alias{estimateTranscriptionalBias}
\title{estimateTranscriptionalBias
transcriptional bias estimates as in (Alfonso-Gonzalez, C, et al., 2022)}
\usage{
estimateTranscriptionalBias(promoterDominanceTable, method)
}
\arguments{
\item{promoterDominanceTable}{Table of promoter dominance summaring transcript pairs expression}

\item{method}{option "chisq" or "fisher" test}
}
\value{
table of significantly biased genes
}
\description{
estimateTranscriptionalBias
transcriptional bias estimates as in (Alfonso-Gonzalez, C, et al., 2022)
}
