% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{expand_rel_time}
\alias{expand_rel_time}
\title{Expand abbreviated period expression}
\usage{
expand_rel_time(abbr)
}
\arguments{
\item{abbr}{The abbreviated period expression}
}
\value{
A list with the sign, value, unit, expanded unit and expression
}
\description{
Expand a shorthand abbreviation like "-1m" to a list with a sign, value and unit
}
