% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{file_modified_less_than}
\alias{file_modified_less_than}
\title{Check if file is recent}
\usage{
file_modified_less_than(fpath, interval = 60)
}
\arguments{
\item{fpath}{The file to check}

\item{interval}{The interval in seconds the file may have changed within}
}
\value{
Logical (TRUE) if file has recently changed
}
\description{
Check if file was modified less than interval seconds
}
