% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{floor_date_extended}
\alias{floor_date_extended}
\title{Extended options for the floor_date function}
\usage{
floor_date_extended(x, unit = c("second", "minute", "hour", "Workday",
  "quarter", "day", "week", "Month", "year"), workday_bounds = c(8, 20))
}
\arguments{
\item{x}{The datetime object}

\item{unit}{The extended unit (including workday) to pass in floor_date}

\item{workday_bounds}{The workday time-of-day bounds to use}
}
\value{
A datetime object
}
\description{
Currently just added Workday
}
