% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{fmt_range_seconds}
\alias{fmt_range_seconds}
\title{Range formatter - seconds}
\usage{
fmt_range_seconds(r)
}
\arguments{
\item{r}{The range list with datetime objects}
}
\value{
The new range with integers expressing the seconds
}
\description{
Format timestamps in range to integer seconds
}
