% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{from_seconds}
\alias{from_seconds}
\title{as.POSIXct wrapper}
\usage{
from_seconds(x)
}
\arguments{
\item{x}{The integer which stores the seconds}
}
\value{
A POSIXct object
}
\description{
as.POSIXct wrapper to convert seconds since epoch to datetime class
}
