% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is_directory}
\alias{is_directory}
\title{Check if input is directory}
\usage{
is_directory(x)
}
\arguments{
\item{x}{The name of the file system object to check}
}
\value{
Logical (TRUE) if x is a directory
}
\description{
Check if input is a directory
}
