% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is_readable}
\alias{is_readable}
\title{Check if readable}
\usage{
is_readable(x)
}
\arguments{
\item{x}{The name of the file system object to check}
}
\value{
Logical (TRUE) if x is readable
}
\description{
Check if file system object x is readable
}
