% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoform_links_database-class.R
\name{isoform_database<-,IsoformDatabase-method}
\alias{isoform_database<-,IsoformDatabase-method}
\title{isoform_database}
\usage{
\S4method{isoform_database}{IsoformDatabase}(x) <- value
}
\arguments{
\item{IsoformDatabase}{}
}
\description{
isoform_database
}
