% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoform_links_database-class.R
\docType{class}
\name{IsoformDatabase-class}
\alias{IsoformDatabase-class}
\title{S4 class of promoter database.}
\description{
S4 class of promoter database.
}
\section{Slots}{

\describe{
\item{\code{isoform_database}}{data.frame.}

\item{\code{TESCoordinate.bins}}{GRanges.}

\item{\code{TESCoordinate.bas}}{GRanges.}

\item{\code{TSSCoordinate.bins}}{GRanges.}

\item{\code{TSSCoordinate.base}}{GRanges.}
}}

