% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoform_links_database-class.R
\name{IsoformDatabase}
\alias{IsoformDatabase}
\title{Build isoform database object}
\usage{
IsoformDatabase(
  isoform_database = data.frame(),
  TESCoordinate.bins = GenomicRanges::GRanges(),
  TESCoordinate.base = GenomicRanges::GRanges(),
  TSSCoordinate.bins = GenomicRanges::GRanges(),
  TSSCoordinate.base = GenomicRanges::GRanges()
)
}
\arguments{
\item{TSSCoordinate.base}{}
}
\description{
Build isoform database object
}
