% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/late.R
\name{late}
\alias{late}
\title{Late runner}
\usage{
late(schedules, retriever, batch_process, forward_process, ...)
}
\arguments{
\item{retriever}{The function which is going to retrieve new data}

\item{batch_process}{The function responsible for processing this batch}

\item{forward_process}{The function which will handle post-processing tasks}

\item{...}{Extra arguments to be passed to the handler}
}
\description{
Run an analytical pipeline in scheduled batches
}
\examples{
# Call crond with late() as the main handler and 
# passing extra arguments (late handlers)
crond(c("5m:10s"), my_tasks, "callback", "late", 
      "retrieve_dt", "batch_process", "forward")

}
