% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LATER-class.R
\docType{class}
\name{LATER-class}
\alias{LATER-class}
\title{S4 class of promoter database.}
\description{
S4 class of promoter database.
}
\section{Slots}{

\describe{
\item{\code{result}}{data.frame.}

\item{\code{stats}}{data.frame.}

\item{\code{readAssignments}}{data.frame.}

\item{\code{isoformCounts}}{data.frame.}
}}

