% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/later-package.R
\docType{package}
\name{later-package}
\alias{later-package}
\title{later: Utilities for Scheduling Functions to Execute Later with Event Loops}
\description{
Executes arbitrary R or C functions some time after the current time, after the R execution stack has emptied. The functions are scheduled in an event loop.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r-lib.github.io/later/}
  \item \url{https://github.com/r-lib/later}
  \item Report bugs at \url{https://github.com/r-lib/later/issues}
}

}
\author{
\strong{Maintainer}: Charlie Gao \email{charlie.gao@posit.co} (\href{https://orcid.org/0000-0002-0750-061X}{ORCID})

Authors:
\itemize{
  \item Winston Chang \email{winston@posit.co}
  \item Joe Cheng \email{joe@posit.co}
}

Other contributors:
\itemize{
  \item Posit Software, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
  \item Marcus Geelnard (TinyCThread library, https://tinycthread.github.io/) [contributor, copyright holder]
  \item Evan Nemerson (TinyCThread library, https://tinycthread.github.io/) [contributor, copyright holder]
}

}
\keyword{internal}
