% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/later.R
\docType{package}
\name{later}
\alias{later}
\alias{later-package}
\title{later}
\description{
A Periodic Batch Processing Framework in R
}
\details{
This is a package that aids in running analytical procedures in a periodic
schedule. It may be useful in batch processing pipelines where new data are
made available periodically.
}
