% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/later.R
\name{loop_empty}
\alias{loop_empty}
\title{Check if later loop is empty}
\usage{
loop_empty()
}
\description{
Returns true if there are currently no callbacks that are scheduled to
execute in the present or future.
}
\keyword{internal}
