% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/later.R
\name{loop_queue_length}
\alias{loop_queue_length}
\title{Check how many callbacks are queued in a later loop}
\usage{
loop_queue_length(loop = current_loop())
}
\arguments{
\item{loop}{A handle to an event loop.}
}
\description{
Returns the number of callbacks that are scheduled to execute in the present
or future.
}
