% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{make_fun_kill_zombies}
\alias{make_fun_kill_zombies}
\title{Zombie processes killer-function generator}
\usage{
make_fun_kill_zombies()
}
\description{
Return a C function compiled with the inline package to kill zombie processes
ref: http://stackoverflow.com/a/25393438/1305020
}
