% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{model_load}
\alias{model_load}
\title{Load the model from a file using readRDS robust}
\usage{
model_load(fpath)
}
\arguments{
\item{fpath}{the file path to the saved model}
}
\value{
the model
}
\description{
Load the model from a file using readRDS robust
}
