% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{model_save}
\alias{model_save}
\title{Save the model to a file using a robust function}
\usage{
model_save(dt_model, fpath, backup_on_overwrite = TRUE)
}
\arguments{
\item{dt_model}{The model to save}

\item{fpath}{Where to save it}

\item{backup_on_overwrite}{Whether to keep a backup on overwrite}
}
\value{
invisible(NULL)
}
\description{
Save the model to a file using a robust function
}
