% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{period_extended}
\alias{period_extended}
\title{Extended period function}
\usage{
period_extended(num = NULL, units = "second", ..., workday_bounds = c(8,
  20))
}
\arguments{
\item{num}{How many units of that period}

\item{units}{Which time unit}

\item{...}{The dots are passed to period}

\item{workday_bounds}{The workday time-of-day bounds to use}
}
\value{
A period object
}
\description{
Currently just added Workday
}
