% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LATER-plot.R
\name{plotTranscriptionalBias}
\alias{plotTranscriptionalBias}
\title{plotTranscriptionalBias}
\usage{
plotTranscriptionalBias(LATER, xlim, ylim, residual, showGenes)
}
\arguments{
\item{showGenes}{}
}
\description{
plotTranscriptionalBias
}
