% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countLinks.R
\name{prepareForCountEnds}
\alias{prepareForCountEnds}
\title{prepare counts trimming the reads to their most 3' into a 1 nt window for counting}
\usage{
prepareForCountEnds(x, window)
}
\arguments{
\item{x}{alignments}

\item{window}{window for trimming}
}
\value{

}
\description{
prepare counts trimming the reads to their most 3' into a 1 nt window for counting
}
