% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoform_links_database-helper.R
\name{prepareIsoformDatabase}
\alias{prepareIsoformDatabase}
\title{prepareIsoformDatabase
Create a reference database of all TSS-TES link associations based on annotation}
\usage{
prepareIsoformDatabase(annotation, tss.window, tes.window)
}
\arguments{
\item{annotation}{reference annotation GenomicRanges object}

\item{tss.window}{window to consider distinct TSS}

\item{tes.window}{window to consider distinct TESassemblies}
}
\value{
promoterDatabase with classification of links
}
\description{
prepareIsoformDatabase
Create a reference database of all TSS-TES link associations based on annotation
}
