% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareLinksDatabase.R
\name{prepareLinksDatabase}
\alias{prepareLinksDatabase}
\title{prepareLinksDatabase
Create a reference database of all TSS-TES link associations based on annotation}
\usage{
prepareLinksDatabase(annotation, tss.window, tes.window)
}
\arguments{
\item{annotation}{reference annotation GenomicRanges object}

\item{tss.window}{window to consider distinct TSS}

\item{tes.window}{window to consider distinct TES}
}
\value{
promoterDatabase with classification of links
}
\description{
prepareLinksDatabase
Create a reference database of all TSS-TES link associations based on annotation
}
