% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{range_recent_tc}
\alias{range_recent_tc}
\title{Construct Ranges of datetimes}
\usage{
range_recent_tc(tc = c("hour", "sec", "min", "Workday", "day", "week",
  "Month", "quarter", "year"), N = 24, what = c("recent", "seasonal"),
  sc = c("day", "hour", "Workday", "week", "Month", "quarter", "year"),
  complete = TRUE, tz = "UTC", anchor, right_open = TRUE)
}
\arguments{
\item{tc}{The primary period to use (for recent time)}

\item{N}{How many units of that period}

\item{what}{Whether to produce a seasonal range sequence or "recent" period range}

\item{sc}{The seasonal period (only used if what = "seasonal")}

\item{complete}{Whether to truncate the reference time to the start of the period (i.e. it is complete)}

\item{tz}{Optionally define a custom timezone}

\item{anchor}{The reference time to use (now by default)}
}
\value{
a single range (recent period) or sequence of ranges (seasonal)
}
\description{
Construct Ranges of datetimes
}
