% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{ranges_from_dates}
\alias{ranges_from_dates}
\title{Ranges from datetime vector}
\usage{
ranges_from_dates(timevector, tc = "hour")
}
\arguments{
\item{timevector}{The vector of datetime objects}

\item{tc}{the time unit to use for constructing the interval}
}
\value{
The list of ranges
}
\description{
Get a list of ranges out of a vector with POSIXct objects
}
