% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{readRDS_robust}
\alias{readRDS_robust}
\title{Robust readRDS}
\usage{
readRDS_robust(fpath, restore_from_backup = TRUE)
}
\arguments{
\item{fpath}{Where to get the rds file from}

\item{restore_from_backup}{Whether to restore from backup and replace a corrupt file}
}
\value{
The object or NULL if the read fails
}
\description{
Read an RDS file with some checking and restoring from a backup
}
