% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countLinks.R
\name{readTESassignment}
\alias{readTESassignment}
\title{Title}
\usage{
readTESassignment(endsAlignements, TESCoordinate.base)
}
\arguments{
\item{endsAlignements}{trimmed reads  output from prepareForCountEnds}

\item{TESCoordinate.base}{reference database for TES}
}
\description{
Title
}
