% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{remove_lock}
\alias{remove_lock}
\title{Remove lock file}
\usage{
remove_lock(lock_file = "service.lock")
}
\arguments{
\item{lock_file}{The path to the lock file}
}
\value{
The return value of file.remove
}
\description{
Companion to add_lock and wait_lock
}
