% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scope.R
\name{scope_dir}
\alias{scope_dir}
\title{Scope to Directory}
\usage{
scope_dir(directory)
}
\arguments{
\item{directory}{The working directory to use.}
}
\description{
Sets the working directory, and resets it at the end
of the active scope.
}

