% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scope.R
\name{scope_env_vars}
\alias{scope_env_vars}
\title{Scope with Environment Variables}
\usage{
scope_env_vars(...)
}
\arguments{
\item{...}{Named arguments, mapping environment variable
names to values.}
}
\description{
Set environment variables for the current scope, and resets it
at the end of the active scope.
}

