% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scope.R
\name{scope_libpaths}
\alias{scope_libpaths}
\title{Scope with Library Paths}
\usage{
scope_libpaths(paths)
}
\arguments{
\item{paths}{The library paths to set.}
}
\description{
Set the library paths (through \code{.libPaths()}) for
the duration of the calling function.
}

