% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scope.R
\name{scope_locale}
\alias{scope_locale}
\title{Scope with Locale}
\usage{
scope_locale(category = "LC_COLLATE", locale)
}
\arguments{
\item{category}{The locale aspect, as a character string.}

\item{locale}{The locale name.}
}
\description{
Set aspects of the current locale and restore the original locale at the
end of the active scope. See \code{\link{Sys.getlocale}()} and
\code{\link{Sys.setlocale}()} for more information.
}
\seealso{
Other scope-related functions: \code{\link{defer}},
  \code{\link{scope_dir}}, \code{\link{scope_env_vars}},
  \code{\link{scope_options}}, \code{\link{scope_path}}
}

