% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scope.R
\name{scope_options}
\alias{scope_options}
\title{Scope with R Options}
\usage{
scope_options(...)
}
\arguments{
\item{...}{Named arguments, mapping option names to values.}
}
\description{
Set \R options (through \code{options}) and restore at the
end of the active scope.
}

