% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scope.R
\name{scope_path}
\alias{scope_path}
\title{Scope with PATH}
\usage{
scope_path(path)
}
\arguments{
\item{path}{A vector of paths.}
}
\description{
Set the \code{PATH} environment variable for the duration
of the active scope.
}
\seealso{
Other scope-related functions: \code{\link{defer}},
  \code{\link{scope_dir}}, \code{\link{scope_env_vars}},
  \code{\link{scope_locale}}, \code{\link{scope_options}}
}

