% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{sd_update}
\alias{sd_update}
\title{Incremental Standard deviation calculation}
\usage{
sd_update(stdev, avg, cnt, val, na.rm = TRUE)
}
\arguments{
\item{stdev}{The current standard deviation}

\item{avg}{The current average value}

\item{cnt}{The number of elements before the new value}

\item{val}{The new value}

\item{na.rm}{Whether to remove NA values}
}
\value{
The new standard deviation
}
\description{
Online algorithm for updating a given std when a new value is added
http://stats.stackexchange.com/questions/24878/computation-of-new-standard-deviation-using-old-standard-deviation-after-change
Algorithms for calculating variance : http://en.wikipedia.org/wiki/Algorithms_for_calculating_variance
TODO: check for numerical stability
}
