% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoform_links_database-class.R
\name{showGene,IsoformDatabase-method}
\alias{showGene,IsoformDatabase-method}
\title{sshowgene}
\usage{
\S4method{showGene}{IsoformDatabase}(x, GENE_ID)
}
\arguments{
\item{IsoformDatabase}{}
}
\value{
IsoformDatabase
}
\description{
sshowgene
}
