% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoform_links_database-class.R
\name{showLinks<-,IsoformDatabase-method}
\alias{showLinks<-,IsoformDatabase-method}
\title{isoform_database}
\usage{
\S4method{showLinks}{IsoformDatabase}(x) <- value
}
\arguments{
\item{IsoformDatabase}{}
}
\description{
isoform_database
}
