% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signals.R
\name{signals}
\alias{off}
\alias{on}
\alias{signal}
\alias{signals}
\title{Signals}
\usage{
on(signal, fn, scoped = TRUE)

off(signal, id)

signal(signal, ...)
}
\arguments{
\item{signal}{The name of a signal, as a string.}

\item{fn}{A function to be executed in response to a signal.}

\item{scoped}{Boolean; should this signal handler be scoped
to the current function body?}

\item{id}{A signal id, as returned by \code{on}.}

\item{...}{Arguments to be passed to registered signal handlers.}
}
\description{
Simple tools for emitting signals (events).
}
\examples{
# Show how 'on', 'signal' can be used to execute
# R functions as requested
counter <- 0
foo <- function() {
  on("increment", function() { counter <<- counter + 1})
  signal("increment")
}

# Call once -- increment counter to 1
foo()
print(counter)

# Call again -- increment counter to 2
foo()
print(counter)

# Fire signal with no listener -- no increment
signal("increment")
print(counter)

}

