% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{split_data_on_field}
\alias{split_data_on_field}
\title{Split dataset on factor and saved parts to files}
\usage{
split_data_on_field(DT, data_path = "data/cache", split_field)
}
\arguments{
\item{DT}{The data frame to split}

\item{data_path}{The data path to save the parts of the dataset}

\item{split_field}{The field name to use for the splitting}
}
\value{
TRUE if successful
}
\description{
Split dataset DT on subfolders defined by the distinct values of split_field
}
