% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{split_data_on_field_and_date}
\alias{split_data_on_field_and_date}
\title{Split and save dataset by date and field}
\usage{
split_data_on_field_and_date(DT, data_path = "data/cache", split_field,
  date_field, date_unit = c("hour", "day", "Month", "year", "week", "minute"))
}
\arguments{
\item{DT}{The data table/frame to split}

\item{data_path}{Where to save the split parts}

\item{split_field}{Which field to use for splitting}

\item{date_field}{The date field to use for date splitting}

\item{date_unit}{The unit to round dates to}
}
\value{
invisible(NULL)
}
\description{
Split dataset DT on both a timestamp field and a categorical field
}
