% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{task_add}
\alias{task_add}
\title{Add a task to TaskList}
\usage{
task_add(TaskList, schedule, FUN, args, method = "fork")
}
\arguments{
\item{TaskList}{An existing list to add tasks to}

\item{schedule}{A shorthand period to recurrently run this task}

\item{FUN}{The function to run}

\item{args}{Optional arguments to FUN}

\item{method}{Whether to run in the parent process or in a parallel R process}
}
\value{
The updated TaskList list
}
\description{
Add a task to our Tasks structure
}
