% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{task_change}
\alias{task_change}
\title{Change a task in a TaskList}
\usage{
task_change(task_id, TaskList, schedule, FUN, args, method)
}
\arguments{
\item{task_id}{The id of the task you want to change}

\item{TaskList}{An existing TaskList list}

\item{schedule}{A shorthand period to recurrently run this task}

\item{FUN}{The function to run}

\item{args}{Optional arguments to FUN}

\item{method}{Whether to run in the parent process or in a parallel R process}
}
\value{
The updated TaskList list
}
\description{
Change the definition of a Task and return the updated TaskList
}
