% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{task_handler}
\alias{task_handler}
\title{A TaskList handler}
\usage{
task_handler(task_id, timestamp, TasksInfo, ...)
}
\arguments{
\item{task_id}{The id of the task to "handle"}

\item{timestamp}{The timestamp this task was scheduled to run}

\item{TasksInfo}{The list (TaskList) with the task metadata}
}
\value{
Depends on the function run by the task and the method used (internal/fork)
}
\description{
A task handler which supports TaskList lists
}
