% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{task_list_reset}
\alias{task_list_reset}
\title{Reset a task list}
\usage{
task_list_reset(TaskList)
}
\arguments{
\item{TaskList}{The TaskList list to reset}
}
\value{
The empty list
}
\description{
Reset a task list - just return an empty list
}
