% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{task_list_schedules}
\alias{task_list_schedules}
\title{Extract the schedules vector}
\usage{
task_list_schedules(TaskList, fieldName = "schedule")
}
\arguments{
\item{TaskList}{An existing list which has some defined schedules}

\item{fieldName}{The fieldName which holds the schedules}
}
\value{
The vector of schedules
}
\description{
Get the schedules defined in a TaskList
}
