% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoform_links_database-class.R
\name{TESCoordinate.base<-,IsoformDatabase-method}
\alias{TESCoordinate.base<-,IsoformDatabase-method}
\title{Title}
\usage{
\S4method{TESCoordinate.base}{IsoformDatabase}(x) <- value
}
\arguments{
\item{IsoformDatabase}{}
}
\description{
Title
}
