% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoform_links_database-class.R
\name{TESCoordinate.base}
\alias{TESCoordinate.base}
\title{TESCoordinate.base}
\usage{
TESCoordinate.base(x)

TESCoordinate.base(x)
}
\arguments{
\item{x}{}
}
\description{
TESCoordinate.base

TESCoordinate.base
}
