% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoform_links_database-class.R
\name{TESCoordinate.bins,IsoformDatabase-method}
\alias{TESCoordinate.bins,IsoformDatabase-method}
\title{TESCoordinate.base}
\usage{
\S4method{TESCoordinate.bins}{IsoformDatabase}(x)
}
\arguments{
\item{IsoformDatabase}{}
}
\description{
TESCoordinate.base
}
