% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoform_links_database-class.R
\name{TSSCoordinate.base,IsoformDatabase-method}
\alias{TSSCoordinate.base,IsoformDatabase-method}
\title{TSSCoordinate.base}
\usage{
\S4method{TSSCoordinate.base}{IsoformDatabase}(x)
}
\arguments{
\item{IsoformDatabase}{}
}
\description{
TSSCoordinate.base
}
