% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoform_links_database-class.R
\name{TSSCoordinate.bins,IsoformDatabase-method}
\alias{TSSCoordinate.bins,IsoformDatabase-method}
\title{TSSCoordinate.bins}
\usage{
\S4method{TSSCoordinate.bins}{IsoformDatabase}(x)
}
\arguments{
\item{IsoformDatabase}{}
}
\description{
TSSCoordinate.bins
}
