% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{with_lock}
\alias{with_lock}
\title{Run expression with a lock file set}
\usage{
with_lock(expr, lock_file = getOption("lock_file", "service.lock"))
}
\arguments{
\item{expr}{The expression to run}

\item{lock_file}{The path to the lock file}
}
\value{
the return value of the expression
}
\description{
Run expr after adding a lock - finally remove the lock
}
