% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{withErrorControl}
\alias{withErrorControl}
\title{Error recovery and logging wrapper}
\usage{
withErrorControl(expr, silentSuccess = FALSE, stopIsFatal = TRUE)
}
\arguments{
\item{expr}{The expression to run}

\item{silentSuccess}{Whether to hide messages on success}

\item{stopIsFatal}{Whether to stop the process when an error occurs}
}
\value{
Whatever the expression returns
}
\description{
Run expression with error recovery and logging support
}
