% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signals.R
\name{events}
\alias{emit}
\alias{events}
\alias{off}
\alias{on}
\alias{stop_propagation}
\title{events}
\usage{
on(event, fn, scoped = TRUE)

off(event, id)

emit(event, ...)

stop_propagation()
}
\arguments{
\item{event}{The name of a event, as a string.}

\item{fn}{A function to be executed in response to a emit.}

\item{scoped}{Boolean; should this event handler be scoped
to the current function body?}

\item{id}{A handler id, as returned by \code{on}. Use this to
remove a registered handler with \code{off} later.}

\item{...}{Arguments to be passed to registered event handlers.}
}
\description{
Simple tools for emitting events (events).
}
\examples{
# Show how 'on', 'emit' can be used to execute
# R functions as requested
counter <- 0
foo <- function() {
  on("increment", function() { counter <<- counter + 1})
  emit("increment")
}

# Call once -- increment counter to 1
foo()
print(counter)

# Call again -- increment counter to 2
foo()
print(counter)

# Fire emit with no listener -- no increment
emit("increment")
print(counter)

}

