% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signals.R
\name{events}
\alias{emit}
\alias{events}
\alias{off}
\alias{on}
\alias{stop_propagation}
\title{Events}
\usage{
on(event, fn, scoped = TRUE)

off(event, id)

emit(event, ...)

stop_propagation()
}
\arguments{
\item{event}{The name of a event, as a string.}

\item{fn}{A function to be executed in response to a emit.}

\item{scoped}{Boolean; should this event handler be scoped
to the current function body?}

\item{id}{A handler id, as returned by \code{on}. Use this to
remove a registered handler with \code{off} later.}

\item{...}{Arguments to be passed to registered event handlers.}
}
\description{
Simple tools for emitting and handling events. \code{on} and \code{off} are
used to register and unregister event handlers, and \code{emit} is used to
emit an event for registered handlers to handle.
}
\details{
These functions interact with a global handler registration -- if you need to
define separate handler registrations, you can use
\code{\link{create_handler_registration}} to generate your own, and call the
relevant \code{on}, \code{off} and \code{emit} events on that object.
}
\examples{
# use event system in increment a counter
counter <- 0
foo <- function() {
  on("increment", function() { counter <<- counter + 1})
  emit("increment")
}

# call once -- increment counter to 1
foo()
print(counter)

# call again -- increment counter to 2
foo()
print(counter)

# fire event with no listener -- no increment
emit("increment")
print(counter)

# generate a new handler registration, and fire an
# event on that object
registration <- create_handler_registration()
registration$emit("increment")

# counter is still 2
print(counter)

}

