% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LATER-class.R
\name{LATER}
\alias{LATER}
\title{Builder LATER}
\usage{
LATER(
  result = data.frame(),
  stats = data.frame(),
  dominance = data.frame(),
  readAssignments = data.frame(),
  isoformCounts = data.frame()
)
}
\arguments{
\item{isoformCounts}{}
}
\description{
Builder LATER
}
