% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/later.R
\name{later}
\alias{later}
\title{Executes a function later}
\usage{
later(func)
}
\arguments{
\item{func}{A function or formula.}
}
\description{
Executes a function or formula (see \code{\link[rlang]{as_function}}) some
time in the future, when no other R code is on the execution stack.
}
\details{
TODO: Talk about error handling
}
