% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/later.R
\name{later_fd}
\alias{later_fd}
\title{Executes a function when a file descriptor is ready}
\usage{
later_fd(func, fd_set, timeout = Inf, loop = current_loop())
}
\arguments{
\item{func}{A function that takes a single argument, a logical vector that
indicates which file descriptors are ready. This may be all \code{FALSE} if the
\code{timeout} argument is non-\code{Inf}.}

\item{fd_set}{Integer vector of file descriptors to monitor. Or on Windows,
these may be \code{HANDLE} values.}

\item{timeout}{Number of seconds to wait before giving up, and calling \code{func}
with all \code{FALSE}.}

\item{loop}{A handle to an event loop. Defaults to the currently-active loop.}
}
\value{
A function, which, if invoked, will cancel the callback. The
function will return \code{TRUE} if the callback was successfully
cancelled and \code{FALSE} if not (this occurs if the callback has
executed or has been cancelled already).
}
\description{
Executes a function when a file descriptor is ready
}
