% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/later.R
\name{later_fd}
\alias{later_fd}
\title{Executes a function when a file descriptor is ready}
\usage{
later_fd(func, fd_set, timeout = Inf, loop = current_loop())
}
\arguments{
\item{func}{A function that takes a single argument, a logical vector that
indicates which file descriptors are ready. This may be all \code{FALSE} if the
\code{timeout} argument is non-\code{Inf}.}

\item{fd_set}{Integer vector of file descriptors to monitor. Or on Windows,
\code{SOCKET} values.}

\item{timeout}{Number of seconds to wait before giving up, and calling \code{func}
with all \code{FALSE}.}

\item{loop}{A handle to an event loop. Defaults to the currently-active loop.}
}
\value{
A function, which, if invoked, will cancel the callback. The
function will return \code{TRUE} if the callback was successfully
cancelled and \code{FALSE} if not (this occurs if the callback has
executed or has been cancelled already).
}
\description{
Executes a function when a file descriptor is ready
}
\examples{
\dontshow{if (requireNamespace("nanonext", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# create nanonext socket
library(nanonext)
s1 <- socket(listen = "inproc://nano")
s2 <- socket(dial = "inproc://nano")
fd1 <- opt(s1, "recv-fd")
fd2 <- opt(s2, "recv-fd")

# 1. timeout: prints FALSE, FALSE
later_fd(print, c(fd1, fd2), 1)
Sys.sleep(1.1)
run_now()

# 2. fd1 active: prints TRUE, FALSE
later_fd(print, c(fd1, fd2), 1)
send(s2, "msg")
Sys.sleep(0.1)
run_now()

# 3. fd2 active: prints FALSE, TRUE
recv(s1)
send(s1, "msg")
later_fd(print, c(fd1, fd2), 1)
Sys.sleep(0.1)
run_now()

close(s2)
close(s1)
\dontshow{\}) # examplesIf}
}
