% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signals.R
\name{signals}
\alias{on}
\alias{signal}
\alias{signals}
\title{Event Handling}
\usage{
on(signal, fn)

signal(signal, ...)
}
\description{
Event Handling
}
\examples{
local({

  # Show how 'on', 'signal' can be used to execute
  # R functions as requested
  counter <- 0
  foo <- function() {
    on("increment", function() { counter <<- counter + 1})
    signal("increment")
  }

  # Call once -- increment counter to 1
  foo()
  print(counter)

  # Call again -- increment counter to 2
  foo()
  print(counter)

  # Fire signal with no listener -- no increment
  signal("increment")
  print(counter)

})

}

